simple-django-redis/requirements.txt
django-redis==4.12.1
Django==2.2.0
djangorestframework~=3.9.0
mysqlclient==1.3.13